requirements.txt
langchain
langchain_community
langchain_core
langgraph
python-dotenv
langchain-openai
pandas
structlog
pymupdf4llm
streamlit
#pygraphviz
